GAWB-23: Refactor TSVFileSupport
